add: sanitize every single variable going through the backend + remove availability of unused method (DELETE, PATCH, PUT)

+++ signaling/server/signaling-server/user_info.py
@@ -5,7 +5,7 @@ from django.core.serializers import serialize
 from django.http import HttpResponse
 import json
 import os
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseServerError
 
 #
 def verifyUser(userJson):
@@ -16,13 +16,19 @@ def verifyUser(userJson):
 
 def getOrCreateUser(response):
 	try:
+		if utils.parse_input(response['hashLogin']) is not True:
+			return HttpResponseServerError("Error: Forbidden characters in request.\n")
 		user_info = UserInfo.objects.get(hashLogin = response['hashLogin'])
 		if user_info.isAnonymized is True:
 			user_info.login = "Anon"
 		else:
 			if 'login' in response:
 				user_info.login = response['login']
+				if utils.parse_input(response['login']) is not True:
+					return HttpResponseServerError("Error: Forbidden characters in request.\n")
 		user_info.token = response['token']
+		if utils.parse_input(response['token']) is not True:
+			return HttpResponseServerError("Error: Forbidden characters in request.\n")
 		user_info.save()
 	except ObjectDoesNotExist:
 		user_info = UserInfo(
@@ -35,18 +41,22 @@ def getOrCreateUser(response):
 	return serialize('json', [user_info])
 
 def storeUserStatistics(request):
-	request = json.loads(request.body.decode())
-	if verifyUser(request) is True:
-		user_info = UserInfo.objects.get(hashLogin = request['hashLogin'])
-		user_info.ballDistance += request['ballDistance']
-		user_info.ballReceived += request['ballReceived']
-		user_info.ballReturned += request['ballReturned']
-		user_info.bonusTaken += request['bonusTaken']
-		user_info.bonusTotal += request['bonusTotal']
-		user_info.gamesPlayedNb += request['gamesPlayedNb']
-		user_info.gameHistory = request['gameHistory']
-		user_info.loseGameNb += request['loseGameNb']
-		user_info.wonGamesNb += request['wonGamesNb']
+	request = request.body.decode()
+	requestJson = json.loads(request)
+	if utils.sanitizeRequest(requestJson) is not True:
+		return HttpResponseServerError(utils.sanitizeRequest(request))
+
+	if verifyUser(requestJson) is True:
+		user_info = UserInfo.objects.get(hashLogin = requestJson['hashLogin'])
+		user_info.ballDistance += requestJson['ballDistance']
+		user_info.ballReceived += requestJson['ballReceived']
+		user_info.ballReturned += requestJson['ballReturned']
+		user_info.bonusTaken += requestJson['bonusTaken']
+		user_info.bonusTotal += requestJson['bonusTotal']
+		user_info.gamesPlayedNb += requestJson['gamesPlayedNb']
+		user_info.gameHistory = requestJson['gameHistory']
+		user_info.loseGameNb += requestJson['loseGameNb']
+		user_info.wonGamesNb += requestJson['wonGamesNb']
 		user_info.save()
 		return HttpResponse(status = 200)
 	else:
@@ -55,6 +65,10 @@ def storeUserStatistics(request):
 def retrieveUser(request):
 	request = request.body.decode()
 	requestJson = json.loads(request)
+
+	if utils.sanitizeRequest(requestJson) is not True:
+		return HttpResponse(utils.sanitizeRequest(requestJson))
+
 	if verifyUser(requestJson) is True:
 		user = UserInfo.objects.get(hashLogin=requestJson['hashLogin'])
 		response = getOrCreateUser(requestJson)
@@ -67,6 +81,9 @@ def retrieveUser(request):
 
 def unanonymizeUser(request):
 	requestJson = json.loads(request.body.decode())
+	if utils.sanitizeRequest(requestJson) is False:
+		return HttpResponseServerError("Error: Forbidden characters in request.\n")
+
 	if verifyUser(requestJson) is True:
 		user = UserInfo.objects.get(hashLogin=requestJson['hashLogin'])
 		user.isAnonymized = False
@@ -77,6 +94,9 @@ def unanonymizeUser(request):
 
 def anonymizeUser(request):
 	requestJson = json.loads(request.body.decode())
+	if utils.sanitizeRequest(requestJson) is False:
+		return HttpResponseServerError("Error: Forbidden characters in request.\n")
+
 	if verifyUser(requestJson) is True:
 		user = UserInfo.objects.get(hashLogin=requestJson['hashLogin'])
 		user.login = "Anon"
@@ -91,6 +111,8 @@ def anonymizeUser(request):
 def deleteUser(request):
 	request = request.body.decode()
 	requestJson = json.loads(request)
+	if utils.sanitizeRequest(requestJson) is False:
+		return HttpResponseServerError("Error: Forbidden characters in request.\n")
 	if verifyUser(requestJson) is True:
 		user = UserInfo.objects.get(hashLogin=requestJson['hashLogin'])
 		user.delete()
